Move SEXP types out of public API
The items defined in the public header file will be stored
in src/OVAL/probes/SEAP/_sexp-types.h instead, which is a private
header file.
